Refactor ContractDetails to use ContractInfo component for improved organization
- Replaced the ContractDetailsTable with a new ContractInfo component to streamline the display of contract statistics and alerts.
- Moved relevant logic and UI elements from ContractDetails to ContractInfo, enhancing code maintainability and readability.
- Updated import statements and adjusted props to ensure seamless integration of the new component.
- Removed redundant code related to contract details display, promoting a cleaner component structure.
